requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
requests>=2.31.0
google-generativeai>=0.3.2
openai>=1.12.0
python-dotenv>=1.0.1